requirements.txt
Django==5.0.6
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.1.0
docker
PyYAML==6.0.1
django-crispy-forms
crispy-bootstrap5
django-jazzmin==2.6.0
whitenoise==6.6.0